Revamped UI, discarded themes.
Changed the way you select a number when playing, with a cool paint(the
program)-styled interface, except on the bottom of the window.

Added support for custom logo even in the taskbar.

Themes are currently disabled, because it looks very much complicated and I
do very much not want to put hours of work into an unimportant feature.