Added checks to prevent crash from inexistent db
The program will now check if the "locale" folder, which is used to
store the database file, exists, and if not will create it.

Also removed "IF EXISTS" check when using the database.py as program
entry point, so that if an inexistent table name is entered, the program
will immediately crash to tell you. This is not problematic, as it is a
very small CLI.